+++ inbox/compile.js
@@ -4,7 +4,7 @@ const solc = require('solc');
 
 /* reading sol file rather than importing it because in import statement node
   expect a .js file but our file is .sol*/
-const inboxPath = path.resolve(__dirname, 'contracts', 'Inbox.sol');
+const inboxPath = path.resolve(__dirname, 'contracts', 'LineOfCredit.sol');
 const source = fs.readFileSync(inboxPath, 'utf8');
 
-module.exports = solc.compile(source, 1).contracts[':Inbox'];
+module.exports = solc.compile(source, 1).contracts[':LineOfCredit'];
